requirements.txt
matplotlib
numpy
torch
torch-geometric
opencv-python
torchvision
numba
tqdm
Pillow
scipy